Move CSR-related code into new certification_request module
